requirements.txt
requests==2.28.0
beautifulsoup4==4.11.1
pytest==7.1.2
requests-mock==1.9.3
numpy
faiss-cpu
openai